src/main/java/org/springframework/samples/travel/interfaces/web/controller/BookingController.java
package org.springframework.samples.travel.interfaces.web.controller;

import java.security.Principal;
import java.util.List;

import javax.inject.Inject;

import org.springframework.samples.travel.application.BookingService;
import org.springframework.samples.travel.domain.model.booking.Booking;
import org.springframework.samples.travel.domain.model.booking.Hotel;
import org.springframework.samples.travel.domain.shared.SearchCriteria;
import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@Controller
public class BookingController {

    private final BookingService bookingService;

    @Inject
    public BookingController(BookingService bookingService) {
        this.bookingService = bookingService;
    }

    @RequestMapping(value = "/hotels/search", method = RequestMethod.GET)
    public void search(SearchCriteria searchCriteria, Principal currentUser, Model model) {
        if (currentUser != null) {
            List<Booking> bookingList = bookingService.findBookings(currentUser.getName());
            model.addAttribute("bookingList", bookingList);
        }
    }

    @RequestMapping(value = "/hotels", method = RequestMethod.GET)
    public String list(SearchCriteria criteria, Model model) {
        List<Hotel> hotels = bookingService.findHotels(criteria);
        model.addAttribute(hotels);
        return "hotels/list";
    }

    @RequestMapping(value = "/hotels/{id}", method = RequestMethod.GET)
    public String show(@PathVariable String id, Model model) {
        model.addAttribute(bookingService.findHotel(id));
        return "hotels/show";
    }

    @RequestMapping(value = "/bookings/{id}", method = RequestMethod.DELETE)
    public String deleteBooking(@PathVariable String id) {
        bookingService.cancelBooking(id);
        return "redirect:../hotels/search";
    }

}